refactor: SM-71 임시 아이템을 Album 컴포넌트로 교체 및 스타일 수정
