+++ src/components/domain/IngredientCarousel/styled.ts
